+++ MissTortas/MissTortasEngine/Model/Product/ProductBase.cs
@@ -3,7 +3,7 @@
     public abstract class ProductBase
     {
         public abstract string Name { get; set; } 
-        public abstract int Id { get; set; }
+        public abstract long Id { get; set; }
         public abstract string Description { get; set; }
     }
 }
